fix: rename module to document_compressors
